refactor: 관심사와 의존성에 따른 패키지 분리
